+++ src/view.c
@@ -109,11 +109,6 @@ void insert_feedback_destroy(struct window_node *node)
     }
 }
 
-static inline CGPoint area_center(struct area a)
-{
-    return (CGPoint) { a.x + a.w*0.5f, a.y + a.h*0.5f };
-}
-
 static inline struct area area_from_cgrect(CGRect rect)
 {
     return (struct area) { rect.origin.x, rect.origin.y, rect.size.width, rect.size.height };
@@ -492,46 +487,74 @@ struct window_node *view_find_min_depth_leaf_node(struct window_node *node)
     return NULL;
 }
 
+static inline bool area_is_in_direction(struct area *r1, CGPoint r1_max, struct area *r2, CGPoint r2_max, int direction)
+{
+    if (direction == DIR_NORTH && r1_max.y <= r2->y) return false;
+    if (direction == DIR_EAST  && r2_max.x <  r1->x) return false;
+    if (direction == DIR_SOUTH && r2_max.y <  r1->y) return false;
+    if (direction == DIR_WEST  && r1_max.x <= r2->x) return false;
+
+    if (direction == DIR_NORTH || direction == DIR_SOUTH) {
+        return ((r2_max.x >= r1->x && r2_max.x <= r1_max.x) ||
+                (r2->x    <= r1->x && r2_max.x >= r1_max.x) ||
+                (r2->x    >= r1->x && r2->x    <= r1_max.x));
+    }
+
+    if (direction == DIR_EAST || direction == DIR_WEST) {
+        return ((r2_max.y >= r1->y && r2_max.y <= r1_max.y) ||
+                (r2->y    <= r1->y && r2_max.y >= r1_max.y) ||
+                (r2->y    >= r1->y && r2->y    <= r1_max.y));
+    }
+
+    return false;
+}
+
+static inline int area_distance_in_direction(struct area *r1, CGPoint r1_max, struct area *r2, CGPoint r2_max, int direction)
+{
+    switch (direction) {
+    case DIR_NORTH: {
+        return r2_max.y > r1->y ? r2_max.y - r1->y : r1->y - r2_max.y;
+    } break;
+    case DIR_EAST: {
+        return r2->x < r1_max.x ? r1_max.x - r2->x : r2->x - r1_max.x;
+    } break;
+    case DIR_SOUTH: {
+        return r2->y < r1_max.y ? r1_max.y - r2->y : r2->y - r1_max.y;
+    } break;
+    case DIR_WEST: {
+        return r2_max.x > r1->x ? r2_max.x - r1->x : r1->x - r2_max.x;
+    } break;
+    }
+
+    return INT_MAX;
+}
+
 struct window_node *view_find_window_node_in_direction(struct view *view, struct window_node *source, int direction)
 {
+    int window_count;
+    uint32_t *window_list = space_window_list(view->sid, &window_count, false);
+    if (!window_list) return NULL;
+
     int best_distance = INT_MAX;
+    int best_rank = INT_MAX;
     struct window_node *best_node = NULL;
-    CGPoint source_point = area_center(source->area);
+
+    CGPoint source_area_max = { source->area.x + source->area.w, source->area.y + source->area.h };
 
     struct window_node *target = window_node_find_first_leaf(view->root);
     while (target) {
-        CGPoint target_point = area_center(target->area);
-        int distance = euclidean_distance(source_point, target_point);
-        if (distance >= best_distance) goto next;
+        CGPoint target_area_max = { target->area.x + target->area.w, target->area.y + target->area.h };
 
-        switch (direction) {
-        case DIR_EAST: {
-            if (target->area.x >= source->area.x + source->area.w) {
+        if (area_is_in_direction(&source->area, source_area_max, &target->area, target_area_max, direction)) {
+            int distance = area_distance_in_direction(&source->area, source_area_max, &target->area, target_area_max, direction);
+            int rank = window_manager_find_rank_of_window_in_list(target->window_order[0], window_list, window_count);
+            if ((distance < best_distance) || (distance == best_distance && rank < best_rank)) {
                 best_node = target;
                 best_distance = distance;
+                best_rank = rank;
             }
-        } break;
-        case DIR_SOUTH: {
-            if (target->area.y >= source->area.y + source->area.h) {
-                best_node = target;
-                best_distance = distance;
-            }
-        } break;
-        case DIR_WEST: {
-            if (target->area.x + target->area.w <= source->area.x) {
-                best_node = target;
-                best_distance = distance;
-            }
-        } break;
-        case DIR_NORTH: {
-            if (target->area.y + target->area.h <= source->area.y) {
-                best_node = target;
-                best_distance = distance;
-            }
-        } break;
         }
 
-next:
         target = window_node_find_next_leaf(target);
     }